Move shape/theme/type to app-shared
- create platform-specific color scheme functions
