+++ app-shared/src/androidMain/kotlin/org/pointyware/timer/shared/ui/theme/Theme.android.kt
@@ -1,0 +1,21 @@
+package org.pointyware.timer.shared.ui.theme
+
+import android.os.Build
+import androidx.compose.material3.ColorScheme
+import androidx.compose.material3.dynamicDarkColorScheme
+import androidx.compose.material3.dynamicLightColorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalContext
+
+@Composable
+actual fun colorScheme(isDarkTheme: Boolean, isDynamicTheme: Boolean): ColorScheme {
+    return if (isDynamicTheme && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+        if (isDarkTheme) {
+            dynamicDarkColorScheme(LocalContext.current)
+        } else {
+            dynamicLightColorScheme(LocalContext.current)
+        }
+    } else {
+        staticColorScheme(isDarkTheme)
+    }
+}
