+++ app-shared/src/commonMain/kotlin/org/pointyware/timer/shared/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.taushsampley.timer.ui.theme
+package org.pointyware.timer.shared.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Shapes
