+++ app-shared/src/commonMain/kotlin/org/pointyware/timer/shared/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.taushsampley.timer.ui.theme
+package org.pointyware.timer.shared.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
