+++ app-shared/src/desktopMain/kotlin/org/pointyware/timer/shared/ui/theme/Theme.desktop.kt
@@ -1,0 +1,7 @@
+package org.pointyware.timer.shared.ui.theme
+
+import androidx.compose.material3.ColorScheme
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun colorScheme(isDarkTheme: Boolean, isDynamicTheme: Boolean): ColorScheme = staticColorScheme(isDarkTheme)
