+++ app-shared/src/jsMain/kotlin/org/pointyware/timer/shared/ui/theme/Theme.js.kt
@@ -1,0 +1,6 @@
+package org.pointyware.timer.shared.ui.theme
+
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun colorScheme(isDarkTheme: Boolean, isDynamicTheme: Boolean) = staticColorScheme(isDarkTheme)
